+++ frontend/e2e/helpers.cafe.ts
@@ -463,4 +463,16 @@ export const waitAndRefresh = async (waitFor = 3000) => {
   await t.eval(() => location.reload())
 }
 
+export const refreshUntilElementVisible = async (selector: string, maxRetries=20) => {
+  const element = Selector(selector);
+  const isElementVisible = async () => await element.exists && await element.visible;
+  let retries = 0;
+  while (retries < maxRetries && !(await isElementVisible())) {
+    await t.eval(() => location.reload()); // Reload the page
+    await t.wait(3000);
+    retries++;
+  }
+  return t.scrollIntoView(element)
+}
+
 export default {}